+++ app/src/main/java/o/mysin/simbirsoftappjava/ui/search/main/SearchTypeNumber.kt
@@ -4,5 +4,10 @@ import o.mysin.simbirsoftappjava.R
 
 enum class SearchTypeNumber(val titleId: Int) {
     EVENTS(R.string.tab_name_search_by_events),
-    NKO(R.string.tab_name_search_by_nko)
+    NKO(R.string.tab_name_search_by_nko);
+
+    companion object {
+        fun getTitleIdByIndex(index: Int): Int = values().getOrNull(index)?.titleId ?: 0
+    }
+
 }
