+++ app/src/main/java/o/mysin/simbirsoftappjava/ui/search/nko/SearchNKOViewModel.kt
@@ -1,4 +1,4 @@
-package o.mysin.simbirsoftappjava.ui.search
+package o.mysin.simbirsoftappjava.ui.search.nko
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,7 +8,7 @@ import kotlinx.coroutines.launch
 import o.mysin.simbirsoftappjava.data.Event
 import java.util.Random
 
-class SearchViewModel : ViewModel() {
+class SearchNKOViewModel : ViewModel() {
 
     private val _nkoList: MutableLiveData<List<Event>> = MutableLiveData()
     val nkoList: LiveData<List<Event>>
